chore: remove promotion
