+++ apps/extension/src/entrypoints/options/pages/api-providers/providers-config.tsx
@@ -18,7 +18,6 @@ import { ConfigCard } from '../../components/config-card'
 import AddProviderDialog from './add-provider-dialog'
 import { selectedProviderIdAtom } from './atoms'
 import { ProviderConfigForm } from './provider-config-form'
-import { addProvider } from './utils'
 
 export function ProvidersConfig() {
   return (
@@ -34,7 +33,7 @@ export function ProvidersConfig() {
       description={(
         <>
           {i18n.t('options.apiProviders.description')}
-          <Promotion className="mt-2" />
+          {/* <Promotion className="mt-2" /> */}
         </>
       )}
       className="lg:flex-col"
@@ -44,31 +43,6 @@ export function ProvidersConfig() {
         <ProviderConfigForm />
       </div>
     </ConfigCard>
-  )
-}
-
-function Promotion({ className }: { className?: string }) {
-  const [providersConfig, setProvidersConfig] = useAtom(configFieldsAtomMap.providersConfig)
-  const setSelectedProviderId = useSetAtom(selectedProviderIdAtom)
-
-  const handleAddAI302Provider = async () => {
-    await addProvider('ai302', providersConfig, setProvidersConfig, setSelectedProviderId)
-  }
-
-  return (
-    <div className={cn('flex items-center gap-2 flex-wrap', className)}>
-      <p className="text-sm text-muted-foreground">{i18n.t('options.apiProviders.promotion.ai302.description')}</p>
-      <div className="flex items-center gap-2">
-        <Button asChild size="sm" className="bg-yellow-500 hover:bg-yellow-500/90">
-          <a href="https://share.302.ai/8o2r7P" target="_blank" rel="noreferrer noopener">
-            {i18n.t('options.apiProviders.promotion.ai302.action')}
-          </a>
-        </Button>
-        <Button size="sm" variant="outline" onClick={handleAddAI302Provider} className="text-black dark:text-white">
-          {i18n.t('options.apiProviders.promotion.ai302.addProvider')}
-        </Button>
-      </div>
-    </div>
   )
 }
